requirements.txt
absl-py==2.2.1
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
grpcio==1.71.0
gymnasium==1.1.1
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.2
pybullet==3.2.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==78.1.0
six==1.17.0
stable_baselines3==2.6.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
typing_extensions==4.13.0
tzdata==2025.2
Werkzeug==3.1.3
PyYAML~=6.0.2